multi mapping page (#7)
* remove double initialization of gamepad GPIOs

* have getProfilePinMappings return whole GpioMappingInfos

before it returned just the actions, which was fine, but now we're
adding more stuff to the infos, so return the whole struct rather than
just the GpioAction within it

* define a custom button combo to a GPIO pin

this expands GpioMappingInfo to include dpad and button masks and uses
those masks when setting up the button maps in Gamepad. this allows for
a GPIO pin to trigger any arbitrary set of buttons

* Add a WIP multi mapping page

* have getProfilePinMappings return whole GpioMappingInfos

before it returned just the actions, which was fine, but now we're
adding more stuff to the infos, so return the whole struct rather than
just the GpioAction within it

* define a custom button combo to a GPIO pin

this expands GpioMappingInfo to include dpad and button masks and uses
those masks when setting up the button maps in Gamepad. this allows for
a GPIO pin to trigger any arbitrary set of buttons

* implement get/setPinMappingsV2 which includes the new masks

* Start moving to new payload structure, new endpoint, translation keys

* Set action if mask is set

* Multi-mapping route added to webconfig, start adding customDpadMask

* Revert endpoint test for v2

* fix issue where setPinMappingsV2 was actually ignoring input

* Rename BUTTON_MASK -> BUTTON_MASK_OPTIONS to prepare for multi mapping data

* Removed some unused imports

* Add support for dpad mask

* Add text for assigned to multi mapping on pin mapping page

* Add support for swapping button labels

* Update pins after setting multi pin

* Set correct default in multi mapping, disables selection on non selectable actions